requirements.txt
flask
requests
google-cloud-firestore
google-cloud-secret-manager
